pyproject.toml
[project]
name = "production-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.21.0",
    "pgvector<0.4",
    "psycopg[binary]>=3.3.3",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "sqlalchemy>=2.0.46",
]

[project.scripts]
production-rag-system = "production_rag_system:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
